refactor(crate): add library entrypoint
